requirements.txt
langchain==0.3.25
langchain_community==0.3.24
langchain-openai==0.3.17
openai==1.79.0
python-dotenv==1.0.1
pydantic==2.11.4
beautifulsoup4==4.13.4
lxml==5.4.1
pypdf==5.5.0
wikipedia==1.4.0